+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - OpenAPI users endpoint offer ability to filter by created_after
 - Fix a bug: providing an array of attributes to filter OpenApi data, results in error
 - Fix a bug: unable to manage stocks on new products
+- Fix a bug: unsupported param[] syntax in OpenAPI
 - Updated react-modal to 3.16.1
 - Updated tiptap editor and its dependencies to 2.0.0-beta.204
 - [TODO DEPLOY] `rails fablab:setup:build_accounting_lines`